修改打印等问题
1.修正在段尾RectItem后面输入文本时出错的问题
2.增加纸张方向属性
3.表格单元格增加垂直对齐方式属性
4.打印问题部分修改
5.字号属性类型修改为Single，支持小数的磅值
6.修正同样5号字，打印出来比word大的问题

+++ source/HCStyleMatch.pas
@@ -49,10 +49,10 @@ type
 
   TFontSizeStyleMatch = class(TStyleMatch)
   private
-    FFontSize: Integer;
+    FFontSize: Single;
   public
     function GetMatchStyleNo(const AStyle: THCStyle; const ACurStyleNo: Integer): Integer; override;
-    property FontSize: Integer read FFontSize write FFontSize;
+    property FontSize: Single read FFontSize write FFontSize;
   end;
 
   TColorStyleMatch = class(TStyleMatch)
